# HHH-11495 Reduce Criteria overhead: add implementors cache in MetamodelImpl