Proposal to fix https://hibernate.atlassian.net/browse/HHH-11495 in Hibernate 5.3