@gsmet 2018-06-29
@caristu I added a few more comments but all in all, it looks like a good idea.

Could you put together a test case allowing to reproduce the performance issue so that I can test the benefit of this approach in a benchmark?

You just need to include it in our test case template: https://github.com/hibernate/hibernate-test-case-templates/tree/master/orm/hibernate-orm-5 and push it to JIRA.

Thanks!