@caristu 2018-06-29
I'm reusing the "implementors" variable to retrieve the cached values.

I can use another variable, if you still prefer ti declare the variable just when needed for the other cases, let me know.