@gsmet 2018-06-29
@caristu no it's OK for the variable. Now that I think about it, maybe we should have a simple `getImplementors()` that does a `implementorsCache.computeIfAbsent(...)` and a `doGetImplementors()` that really does the work. This way, we won't compute the map several times if concurrent accesses happen when the value is not yet computed.

This is small tuning, I can also do it while squashing the patch.

Do you think you'll have the time to put together a test case?