@caristu 2018-06-29
@gsmet ok, sounds great

yes, I'm working it, I'll try to provide the test asap