@caristu 2018-06-29
@gsmet I've just pushed the test case

This is the repo: https://github.com/caristu/hibernate-test-case-templates/tree/criteria-overhead
And this is the test case that can be used to measure time differences with and without the fix: https://github.com/caristu/hibernate-test-case-templates/blob/criteria-overhead/orm/hibernate-orm-5/src/test/java/org/hibernate/bugs/CriteriaOverhead.java