@gsmet 2018-07-02
@caristu I squashed your commits and used `computeIfAbsent()` instead.

It indeed makes a very big difference when the model is large. Wondering how we could have ignored it for so long.

Thanks for the patch. I'll ask @yrodiere to take another look before merging.