@caristu 2018-07-02
@gsmet thanks for handling the PR on so short notice. Would this patch be included in next 5.3.2 version?It would be great.