@Sanne 2018-07-03
Verry nice!

What about memory usage, do we need to cap the cache size and/or is there any opportunity to discard entries?