@gsmet 2018-07-03
@Sanne my understanding is that it's naturally capped by the number of entities you have in your application. And even more the ones you are explicitly querying.

So I don't think we will have an issue with this map. Except if a user executes random queries with random entity names but I don't think it makes any sense :).