@Sanne 2018-07-03
Thanks @gsmet , sounds reasonable.

_getImplementors()_ is public JPA API tough. We should make a defensive copy of that array :/