@gsmet 2018-07-03
@Sanne is it? I don't see it exposed with an Eclipse "Open call hierarchy".