@yrodiere 2018-07-03
@gsmet The method implements `org.hibernate.Metamodel#getImplementors`, which is public API. Not JPA, Hibernate-native API. But still public API.

I pushed a few additional commits to address that, and more.

Regarding the patch, it seems to make sense, but I'm not convinced it's really useful in practice. It could even be damageable, for all I know.

The given benchmark has so much variance that I don't see how we could trust it. Here are the results on my machine:

* without the patch:
  ```
  *** Criteria with 1 threads done in 5317 ms ***
  *** Criteria with 2 threads done in 2 ms ***
  *** Criteria with 3 threads done in 1330 ms ***
  *** Criteria with 4 threads done in 952 ms ***
  *** Criteria with 5 threads done in 764 ms ***
  *** Criteria with 6 threads done in 3 ms ***
  *** Criteria with 7 threads done in 636 ms ***
  *** Criteria with 8 threads done in 3 ms ***
  ```
* with patch:
  ```
  *** Criteria with 1 threads done in 3063 ms ***
  *** Criteria with 2 threads done in 1203 ms ***
  *** Criteria with 3 threads done in 770 ms ***
  *** Criteria with 4 threads done in 677 ms ***
  *** Criteria with 5 threads done in 485 ms ***
  *** Criteria with 6 threads done in 395 ms ***
  *** Criteria with 7 threads done in 3 ms ***
  *** Criteria with 8 threads done in 9 ms ***
  ```
* with patch plus an additional commit adding a `clone()` call in `getImplementors(String)` to protect the cache from external changes:
  ```
  *** Criteria with 1 threads done in 4482 ms ***
  *** Criteria with 2 threads done in 1743 ms ***
  *** Criteria with 3 threads done in 1051 ms ***
  *** Criteria with 4 threads done in 861 ms ***
  *** Criteria with 5 threads done in 718 ms ***
  *** Criteria with 6 threads done in 2 ms ***
  *** Criteria with 7 threads done in 441 ms ***
  *** Criteria with 8 threads done in 410 ms ***
  ```

Or is it me? Should I use some specific configuration to make the benchmark more reliable?