@yrodiere 2018-07-03
Also, if we really go for this, I think we should only cache actual hits, at least as a first step. We don't know how people will use this API, they could potentially iterate on all the classes in the classpath, and then we'd end up with a bloated cache.

Maybe we should cache only in cases where the class passed as a parameter actually has persisted subtypes?