@gsmet 2018-07-03
@Sanne I reworked the method with your advises in mind. Could you take a look? Thanks!