@Sanne 2018-07-03
Nice safer approach to not catch on CNFE.

Why did you decide to not cache on `implementors.length == 0` ?