@gsmet 2018-07-03
@Sanne Yoann's idea. It's to avoid caching elements for classes not resolved as entities e.g. it won't be possible to fill in the cache with random classes.