@gsmet 2018-07-03
Done!

@caristu thanks for having pushed hard for this.
@yrodiere @Sanne thanks for the reviews and ideas, great team effort!