@Sanne 2018-07-03
yea, thanks all! A pleasure to see this shaped collaboratively, and big thanks to @caristu for the core idea.